231019 윤하 2차 커밋 (회원탈퇴 막기..)

+++ src/main/java/com/cruise/project_cruise/controller/MypageController.java
@@ -13,7 +13,9 @@ import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.ModelAndView;
 
 import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
 import java.awt.*;
+import java.io.PrintWriter;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
@@ -41,10 +43,10 @@ public class MypageController {
 
         String email = "hchdbsgk@naver.com";
 
-        List<CrewDTO> crewLists = mypageService.getCrews(email); //크루 정보
+        List<CrewDTO> crewLists = mypageService.getCrews(email); //크루 정보 (크루계좌번호 등록된 계좌 삭제할 때 사용)
         List<CrewMemberDTO> crewNumLists = mypageService.getCrewNums(email); //크루맴버의 크루번호
         List<OpenBankDTO> openAccPwd = mypageService.getOpenAccPWd(email); //가상계좌 비밀번호
-        List<OpenBankDTO> accountLists = mypageService.getAccounts(email); //가상계좌정보
+        List<MyAccountDTO> accountLists = mypageService.getAccountList(email); //계좌 전체 조회
         UserDTO userInfo = mypageService.getUserInfo(email); // 로그인한 사용자 정보.이름
 
         ModelAndView mav = new ModelAndView();
@@ -168,7 +170,7 @@ public class MypageController {
         수정 후 뿌려주기 위해 다시 조회
      */
     @PostMapping("/mypage/updateAname")
-    public List<Map<String,Object>> updateAname(@RequestParam("openAccount") String openAccount,@RequestParam("openAname") String openAname) throws Exception{
+    public List<Map<String,Object>> updateAname(@RequestParam("myaccountNum") String myaccountNum,@RequestParam("myaccountName") String myaccountName) throws Exception{
 
         JSONObject jsonObject = new JSONObject();
         JSONArray jsonArray = new JSONArray();
@@ -176,11 +178,11 @@ public class MypageController {
 
         String email = "hchdbsgk@naver.com";
 
-        mypageService.updateAname(openAname,openAccount); //계좌명 수정
-        List<OpenBankDTO> accountLists = mypageService.getAccounts(email); //가상계좌정보
+        mypageService.updateAname(myaccountName,myaccountNum); //계좌명 수정
+        List<MyAccountDTO> accountLists = mypageService.getOneAccount(email,myaccountNum); //가상계좌정보
 
         for(int i=0;i<accountLists.size();i++){
-            hashMap.put("selectAname", accountLists.get(i).getOpen_aname());
+            hashMap.put("selectAname", accountLists.get(i).getMyaccount_name());
 
             jsonObject = new JSONObject(hashMap);
             jsonArray.add(jsonObject);
@@ -193,17 +195,43 @@ public class MypageController {
     }
 
     /*
+        등록된 계좌 삭제 메소드
+        - 모임통장인 경우, 참조키 제약으로 삭제되지 않음
+     */
+    @PostMapping("/mypage/deleteAccount")
+    public void deleteMyaccount(@RequestParam("myaccountNum") String myaccountNum) throws Exception{
+
+        mypageService.deleteMyaccount(myaccountNum);
+
+    }
+
+    /*
         크루 '탈퇴하기' 버튼 누를 때 get방식으로 삭제하는 메소드
      */
     @GetMapping("/mypage/mypage_all_ok")
-    public ModelAndView delCrew(HttpServletRequest request) throws  Exception {
+    public ModelAndView delCrew(HttpServletRequest request, HttpServletResponse response) throws  Exception {
 
-        String email = "hchdbsgk@naver.com";
-        int crewNum = Integer.parseInt(request.getParameter("crewNum"));
-
-        mypageService.deleteCrew(email,crewNum);
+        response.setContentType("text/html; charset=UTF-8");
+        PrintWriter out = response.getWriter();
 
         ModelAndView mav = new ModelAndView();
+
+        String email = "hchdbsgk@naver.com";
+        int crewNum = Integer.parseInt(request.getParameter("crewNum")); //선택한 크루넘
+        String capEmail = mypageService.getOneCaptain(email,crewNum);
+
+        System.out.println("캡틴 이메일-> " + capEmail);
+
+        if(capEmail.equals(email)){
+
+            out.println("<script> alert('선장이어서 크루 탈퇴 못함!'); ");
+            out.println("window.location.href='/mypage/mypage_all';</script>");
+            out.flush();
+
+
+        }else {
+            mypageService.deleteCrew(email,crewNum);
+        }
 
         mav.setViewName("redirect:/mypage/mypage_all");
